[Updated]ReviewsController - change update reviews - change relation type for review and review image - change review delete algo [Created]ReviewImagesController - create new image - delete image - get image by id [Updated] ReviewtagController - add auth endpoints [Updated] User map profiles [Updated] UserController - end resend email code endpoint

+++ CQRS/Handlers/QueryHandlers/ReviewQueryHandlers/DeleteReviewCommandHandler.cs
@@ -1,25 +1,46 @@
 ﻿using amazon_backend.CQRS.Commands.ReviewRequests;
+using amazon_backend.Data.Entity;
 using amazon_backend.Models;
+using amazon_backend.Services.JWTService;
 using MediatR;
 
 namespace amazon_backend.CQRS.Handlers.QueryHandlers.ReviewQueryHandlers
 {
-    public class DeleteReviewCommandHandler : IRequestHandler<DeleteReviewCommandRequest, Result<Guid>>
+    public class DeleteReviewCommandHandler : IRequestHandler<DeleteReviewCommandRequest, Result<string>>
     {
         private readonly IReviewDao _reviewDao;
-        public DeleteReviewCommandHandler(IReviewDao reviewDao)
+        private readonly TokenService _tokenService;
+
+        public DeleteReviewCommandHandler(IReviewDao reviewDao, TokenService tokenService)
         {
             _reviewDao = reviewDao;
+            _tokenService = tokenService;
         }
-        public async Task<Result<Guid>> Handle(DeleteReviewCommandRequest request, CancellationToken cancellationToken)
+        public async Task<Result<string>> Handle(DeleteReviewCommandRequest request, CancellationToken cancellationToken)
         {
-            var reviewId = Guid.Parse(request.reviewId);
-            var result = await _reviewDao.DeleteAsync(reviewId);
-            if (result)
+            var decodeResult = await _tokenService.DecodeTokenFromHeaders();
+            if (!decodeResult.isSuccess)
             {
-                return new(reviewId);
+                return new() { isSuccess = decodeResult.isSuccess, message = decodeResult.message, statusCode = decodeResult.statusCode };
             }
-            else return new("Review not found");
+            User user = decodeResult.data;
+            var reviewId = Guid.Parse(request.reviewId);
+
+            Review? review = await _reviewDao.GetByIdAsync(reviewId);
+            if (review == null)
+            {
+                return new("Not found") { statusCode = 404 };
+            }
+            if (review.UserId == user.Id || user.Role == "Admin")
+            {
+                var result = await _reviewDao.DeleteAsync(reviewId);
+                if (result)
+                {
+                    return new("Ok") { statusCode = 200 };
+                }
+                return new("See server logs") { statusCode = 500 };
+            }
+            return new("Not found") { statusCode = 404 };
         }
     }
 }
